rebase

+++ CalPatRec/src/CalHelixFinder_module.cc
@@ -122,9 +122,9 @@ namespace mu2e {
 
     ChannelID cx, co;
     int       nPlanesPerStation(2);
-    for (int ipl=0; ipl<_tracker->nPlanes(); ipl++) {
+    for (size_t ipl=0; ipl<_tracker->nPlanes(); ipl++) {
       const Plane*  pln = &_tracker->getPlane(ipl);
-      for (int ipn=0; ipn<pln->nPanels(); ipn++) {
+      for (size_t ipn=0; ipn<pln->nPanels(); ipn++) {
 	const Panel* panel = &pln->getPanel(ipn);
 	int face;
 	if (panel->id().getPanel() % 2 == 0) face = 0;
